SPG-866: CG API Key Error Panics

+++ internal/service/cd/cloudprovider/common.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/harness/terraform-provider-harness/internal"
 	"github.com/harness/terraform-provider-harness/internal/service/cd/usagescope"
+	"github.com/harness/terraform-provider-harness/internal/utils"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -30,7 +31,9 @@ func commonCloudProviderSchema() map[string]*schema.Schema {
 func resourceCloudProviderDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	log.Printf("[DEBUG] Deleting cloud provider %s", d.Get("name"))
 	c := meta.(*internal.Session).CDClient
-
+	if c == nil {
+		return diag.Errorf(utils.CDClientAPIKeyError)
+	}
 	id := d.Get("id").(string)
 	err := c.CloudProviderClient.DeleteCloudProvider(id)
